Fix segfault and compiler warnings

+++ libaudio/src/controlEventExtractor.cpp
@@ -68,12 +68,15 @@ struct ControlEventExtractor::Impl {
          // When sustain pedal is pressed, low-frequency resonances
          // increase, raising the low-frequency energy.
          float rmsLowFreq = 0.0f;
-         uint32_t lowFreqBins = static_cast<uint32_t>(200.0f * windowSize / sr);
+         uint32_t lowFreqBins = static_cast<uint32_t>(
+            200.0f * static_cast<float>(windowSize) /
+            static_cast<float>(sr));
 
          for (uint32_t i = 0; i < std::min(lowFreqBins, framesRead); ++i) {
             rmsLowFreq += buffer[i] * buffer[i];
          }
-         rmsLowFreq = std::sqrt(rmsLowFreq / std::max(1u, lowFreqBins));
+         rmsLowFreq =
+            std::sqrt(rmsLowFreq / static_cast<float>(std::max(1u, lowFreqBins)));
 
          // Threshold-based detection.
          // A high RMS energy in the low-frequency range suggests the
